requirements.txt
llama-cpp-python>=0.2.0
pandas>=1.5.0
openpyxl>=3.0.0
pyodbc>=4.0.0
sqlalchemy>=2.0.0
tqdm>=4.60.0
python-dotenv>=1.0.0
pyyaml>=6.0
pyarrow>=14.0.0
fastparquet>=2024.0.0
psutil>=5.9.0
pytest>=8.0.0
pytest-cov>=5.0.0
